backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "mvanet-segmentation-api"
version = "1.0.0"
description = "High-performance image segmentation API using MVANet"
readme = "docs/README.md"
license = {text = "MIT"}
authors = [
    {name = "Development Team"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.68.0,<1.0.0",
    "uvicorn[standard]>=0.15.0,<1.0.0",
    "torch>=1.9.0",
    "torchvision>=0.10.0",
    "pillow>=8.0.0",
    "numpy>=1.20.0",
    "opencv-python>=4.5.0",
    "einops>=0.3.0",
    "timm>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.0",
    "flake8>=3.8",
    "mypy>=0.800",
    "pre-commit>=2.10",
]
test = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "pytest-asyncio>=0.15",
]

[project.urls]
Homepage = "https://github.com/company/mvanet-segmentation-api"
Repository = "https://github.com/company/mvanet-segmentation-api"
Documentation = "https://github.com/company/mvanet-segmentation-api/blob/main/docs/README.md"

[project.scripts]
mvanet-api = "src.api.app:main"